TST: better workaround for NamedTempfile problem on windows
from https://github.com/matplotlib/matplotlib/pull/5278#discussion_r42749345

+++ lib/matplotlib/tests/test_font_manager.py
@@ -31,11 +31,15 @@ def test_font_priority():
 def test_json_serialization():
     # on windows, we can't open a file twice, so save the name and unlink
     # manually...
-    with tempfile.NamedTemporaryFile(delete=False) as temp:
-        name = temp.name
-    json_dump(fontManager, name)
-    copy = json_load(name)
-    os.remove(name)
+    try:
+        name = None
+        with tempfile.NamedTemporaryFile(delete=False) as temp:
+            name = temp.name
+        json_dump(fontManager, name)
+        copy = json_load(name)
+    finally:
+        if name and os.path.exists(name):
+            os.remove(name)
     with warnings.catch_warnings():
         warnings.filterwarnings('ignore', 'findfont: Font family.*not found')
         for prop in ({'family': 'STIXGeneral'},
